requirements.txt
fastapi~=0.111.0
uvicorn[standard]
gunicorn
pydantic~=2.7.2
python-dotenv~=1.0.0
numpy~=1.23.0
slowapi~=0.1.9
requests~=2.31.0
brotli
sqlalchemy~=2.0.30
httpx~=0.27.0
psycopg2-binary
